refactor(): 자판기를 초기 생성하는 객체를 역할에 따라 분리
